新增"处理超时时间"配置项 修改README

+++ handler/const.go
@@ -3,5 +3,5 @@ package handler
 var (
 	DownloadStickerSetCallbackQuery = "DOWNLOAD_STICKERS_SET"
 	QuitQueueCallbackQueryPrefix    = "QUIT_"
-	ProcessTimeout                  = float64(60)
+	ProcessTimeout                  = 60
 )
